Leave notes for next steps...
Since the `Authorization` header controls whether the user is logged in or not,
removing that value will have the effect of logging the user out: "forgetting"
the encoded username and password. We might also need to do something with the
data that comes back from `/whoami`, in case we want to address the user inside
the application.